drivers: loop: Set WQ_HIGHPRI for the workqueue of the loop device
The ZRAM WriteBack feature on Android devices relies on a loop as the
storage backend. Under high system load, a workqueue priority inversion issue
may occur, leading to the inability to complete IO requests in a timely manner.
Therefore, the WQ_HIGHPRI flag is added to the workqueue of the loop device to
alleviate the priority inversion problem as quickly as possible.

Change-Id: I1dc7de0a3eea14ae6b8a09236ccd7e54e53b80fb